It was taking a long time for gulp to build the html and the assets and gulp was simply piping everything in the html and assets folder. Defaulting to using `cp` instead of gulp pipes reduced the build time of the html stage from 48 secs to 46 ms and the assets stage from 53 secs to 2.79 ms.

Before: `gulp build`
```
Default:
[11:32:57] Requiring external module @babel/register
Browserslist: caniuse-lite is outdated. Please run:
npx browserslist@latest --update-db
[11:32:57] Using gulpfile ~/git/districtr/gulpfile.babel.js
[11:32:57] Starting 'build'...
[11:32:57] Starting 'clean'...
[11:32:57] Finished 'clean' after 624 μs
[11:32:57] Starting 'js'...
[11:32:57] Starting 'css'...
[11:32:57] Starting 'html'...
[11:32:57] Starting 'assets'...
[11:32:57] Starting 'deployFiles'...
[11:32:57] Finished 'deployFiles' after 221 ms
[11:33:33] Finished 'css' after 36 s
[11:33:45] Finished 'html' after 48 s
[11:33:50] Finished 'assets' after 53 s
[11:33:53] Finished 'js' after 56 s
[11:33:53] Finished 'build' after 56 s
```
After:
```
[13:16:46] Requiring external module @babel/register
Browserslist: caniuse-lite is outdated. Please run:
npx browserslist@latest --update-db
[13:16:47] Using gulpfile ~/git/districtr/gulpfile.babel.js
[13:16:47] Starting 'build'...
[13:16:47] Starting 'clean'...
[13:16:47] Finished 'clean' after 2.07 ms
[13:16:47] Starting 'mkdir'...
[13:16:47] Finished 'mkdir' after 1.92 ms
[13:16:47] Starting 'js'...
[13:16:47] Starting 'css'...
[13:16:47] Starting 'html'...
[13:16:47] Starting 'assets'...
[13:16:47] Starting 'deployFiles'...
[13:16:47] Finished 'html' after 57 ms
[13:16:47] Finished 'deployFiles' after 158 ms
[13:16:50] Finished 'assets' after 2.75 s
[13:17:23] Finished 'css' after 36 s
[13:17:44] Finished 'js' after 56 s
[13:17:44] Finished 'build' after 56 s
```

Unfortunately, it looks like now the `js` stage of the build is the bottleneck. But, once that is resolved, build times will be cut to 36 seconds. It looks like `bundleView` is synchronous, so making `bundleView` asynchronous would probably help a lot.