@InnovativeInventor 2021-06-27
Update: dev builds are at 1.73 sec. You can do a dev build by running `npm run develop` (for a live/hot reload version) or `npm run devBuild` (for a static build).